requirements.txt
langchain
pypdf
sentence-transformers
openai
chromadb
langchain_community
streamlit_chat
python-dotenv
langchain_openai